Rename betty.plugin.config to betty.plugin.data
