+++ betty/content_provider/content_providers.py
@@ -25,7 +25,7 @@ from betty.locale.localizable.property import LocalizableProperty
 from betty.locale.localize import resolve_localized
 from betty.media_type import MediaType
 from betty.media_type.media_types import PLAIN_TEXT
-from betty.plugin.config.property import PluginManufacturerSequenceProperty
+from betty.plugin.data.property import PluginManufacturerSequenceProperty
 from betty.sample import Size
 from betty.service.factory import DataManufacturable, Manufacturable
 from betty.service.requirement.project import require_project
