+++ betty/copyright_notice/data.py
@@ -10,7 +10,7 @@ from betty.copyright_notice import CopyrightNotice, CopyrightNoticeDefinition
 from betty.data.aggregate.record.object import ObjectDefinition
 from betty.locale.localizable.gettext import _
 from betty.locale.localizable.property import LocalizableProperty
-from betty.plugin.config import HumanFacingPluginDefinitionConfiguration
+from betty.plugin.data import HumanFacingPluginDefinitionConfiguration
 from betty.sample import Sample
 
 if TYPE_CHECKING:
