+++ betty/plugin/data/__init__.py
@@ -1,13 +1,15 @@
 """
-Provide plugin configuration.
+Data types for plugins.
 """
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, final
 
+from betty.collections import MutableResolvedSequence, MutableResolvedSequenceProxy
 from betty.data import Data
+from betty.data.aggregate.collection.sequence import SequenceDefinition
 from betty.data.aggregate.record.object import ObjectDefinition
 from betty.data.aggregate.record.object.property import Optional
 from betty.locale.localizable.gettext import _
@@ -23,6 +25,7 @@ if TYPE_CHECKING:
         ResolvableCountableLocalizable,
         ResolvableLocalizable,
     )
+    from betty.plugin.factory import PluginManufacturer
 
 
 class PluginDefinitionConfiguration[
@@ -100,3 +103,25 @@ class CountableHumanFacingPluginDefinitionConfiguration[
         super().__init__(**kwargs)
         self.label_plural = label_plural
         self.label_countable = label_countable
+
+
+@final
+class PluginManufacturerSequenceDefinition(SequenceDefinition):
+    """
+    Define a sequence of plugin instance configurations.
+    """
+
+    def __init__(
+        self,
+        manufacturer: type[PluginManufacturer],
+        *,
+        label: ResolvableLocalizable | None = None,
+    ):
+        super().__init__(
+            cls=MutableResolvedSequence,
+            factory=lambda values: MutableResolvedSequenceProxy(
+                list(values), value_resolver=manufacturer.resolve
+            ),
+            value=manufacturer,
+            label=manufacturer.type().type().label_plural if label is None else label,
+        )
