--- betty/tests/plugin/test_data.py
@@ -1,6 +1,0 @@
-class TestPluginManufacturerDefinition:
-    pass
-
-
-class TestPluginManufacturerSequenceDefinition:
-    pass
